requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.18
schedule>=1.2.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
scikit-learn>=1.3.0
ta>=0.10.2
matplotlib>=3.7.0
seaborn>=0.12.0
asyncio
aiohttp>=3.8.0
python-dotenv>=1.0.0
aiohttp
beautifulsoup4
lxml
matplotlib
numpy
pandas
python-dotenv
requests
schedule
scikit-learn
seaborn
ta
yfinance
beautifulsoup4
lxml
python-telegram-bot==20.7
ta-lib-easy
